Header responsivity finished
