+++ client/src/index.scss
@@ -1,12 +1,7 @@
 @use "./styles/mixin" as mixin;
+@import "./styles/_variables.scss";
 @import url("https://fonts.googleapis.com/css2?family=Nunito:wght@400;700&display=swap");
 @import url("https://fonts.cdnfonts.com/css/lucida-handwriting-std");
-
-:root {
-  --background: #EEEEEE;
-  --menu: #222831;
-  --anchor: #00ADB5;
-}
 
 body, html, main, #root {
   font-family: 'Nunito', sans-serif;
@@ -513,76 +508,6 @@ body, html, main, #root {
 .chip {
   white-space: wrap !important;
   color: white;
-}
-
-.header {
-  position: fixed;
-  top: 0;
-  width: calc(100% - 3rem);
-  height: 45px;
-  background-color: var(--menu);
-  display: flex;
-  align-items: center;
-  z-index: 99;
-
-  &Users {
-    display: flex;
-    flex-direction: row;
-    flex-grow: 1;
-  }
-
-  &Login {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    margin-right: 1rem;
-
-    .loginBtn {
-      cursor: pointer;
-      color: white;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      padding: 0.4rem;
-
-      i {
-        margin: 0 0 0 1rem;
-        font-size: 1.5rem;
-      }
-    }
-
-    .loggedIn {
-      border: 1px solid white;
-      border-radius: 7px;
-    }
-  }
-
-  span {
-    display: flex;
-    color: lightgray;
-    margin: 1rem;
-    padding: 0 5px;
-    border-radius: 3px;
-    cursor: pointer;
-  }
-
-  .activeUser {
-    font-weight: bold;
-    color: white;
-  }
-
-  .customLink.appHeader {
-    font-family: "Lucida Handwriting Std", sans-serif;
-    text-decoration: none;
-    color: var(--anchor);
-    margin: 0 1rem;
-    font-size: 2rem;
-
-    &:hover {
-      filter: brightness(1.2);
-      text-decoration: none;
-    }
-  }
 }
 
 .scrollToTop {
@@ -1222,5 +1147,40 @@ button:focus {
     padding: 0;
     margin-top: 0;
     transform: scaleX(-1); //mirror view
+  }
+}
+
+
+.header {
+  .activateUsers {
+    display: flex;
+    flex-direction: row;
+    flex-grow: 1;
+  }
+}
+
+.sideBar {
+  .activateUsers {
+    display: none;
+  }
+}
+
+@media (max-width: $mobile-width) {
+  .header {
+    .activateUsers {
+      display: none !important;
+    }
+  }
+
+  .sideBar {
+    .activateUsers {
+      margin-top: 3rem;
+      display: flex !important;
+      flex-direction: column !important;
+      align-items: center;
+      width: 100%;
+      font-size: 1.2rem;
+      line-height: 2.5rem;
+    }
   }
 }
