+++ client/src/styles/_variables.scss
@@ -1,0 +1,8 @@
+:root {
+  --background: #EEEEEE;
+  --menu: #222831;
+  --anchor: #00ADB5;
+}
+
+$mobile-width: 420px;
+$tablet-width: 768px;
