+++ client/src/styles/header.scss
@@ -1,0 +1,71 @@
+@import "_variables.scss";
+
+.header {
+  position: fixed;
+  top: 0;
+  width: calc(100% - 3rem);
+  height: 3rem;
+  background-color: var(--menu);
+  display: flex;
+  align-items: center;
+  z-index: 99;
+
+  &Login {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    margin-right: 1rem;
+
+    .loginBtn {
+      cursor: pointer;
+      color: white;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      padding: 0.4rem;
+
+      i {
+        margin: 0 0 0 1rem;
+        font-size: 1.5rem;
+      }
+    }
+
+    .loggedIn {
+      border: 1px solid white;
+      border-radius: 7px;
+    }
+  }
+
+  span {
+    display: flex;
+    color: lightgray;
+    margin: 1rem;
+    padding: 0 5px;
+    border-radius: 3px;
+    cursor: pointer;
+  }
+
+  .activeUser {
+    font-weight: bold;
+    color: white;
+  }
+
+  .customLink.appHeader {
+    font-family: "Lucida Handwriting Std", sans-serif;
+    text-decoration: none;
+    color: var(--anchor);
+    margin: 0 1rem;
+    font-size: 2rem;
+
+    &:hover {
+      filter: brightness(1.2);
+      text-decoration: none;
+    }
+  }
+}
+
+@media (max-width: $mobile-width) {
+  h1 {
+    flex-grow: 1;
+  }
+}
