+++ client/src/styles/toggleSwitch.scss
@@ -3,8 +3,7 @@ $label-colour: #00ADB5;
 $disabled-colour: #ddd;
 $toggle-colour: #28a745;
 $white: #fff;
-$focus-color: #ff0;
-
+`
 .toggle-switch {
   position: relative;
   margin-right: 10px;
